Common LocalBind range parsing
