+++ pkg/ast_processing/find_position.go
@@ -1,4 +1,4 @@
-package processing
+package ast_processing
 
 import (
 	"errors"
