+++ pkg/ast_processing/top_level_objects.go
@@ -1,4 +1,4 @@
-package processing
+package ast_processing
 
 import (
 	"github.com/google/go-jsonnet"
